update integration test
