+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 import { generateChainId } from './generateChainId';
 import { sanitizePrivateKey } from './sanitizePrivateKey';
+import { withFallbackPrivateKey } from './withFallbackPrivateKey';
 
-export { generateChainId, sanitizePrivateKey };
+export { generateChainId, sanitizePrivateKey, withFallbackPrivateKey };
